app/requirements.txt
fastapi
uvicorn
pymupdf
weaviate-client
python-dotenv
openai